+++ src/nodes/Core/X3DNode.js
@@ -738,7 +738,7 @@ x3dom.registerNodeType(
                     } );
                 }
 
-                if ( this._vf[ name ] &&
+                if ( this._vf[ name ] !== undefined &&
                     !xmlNode.attributes[ name ] && !xmlNode.attributes[ name.toLowerCase() ] )
                 {
                     var str = "";