+++ src/nodes/H-Anim/HAnimHumanoid.js
@@ -272,6 +272,8 @@ x3dom.registerNodeType(
                 if ( singlePath && ( invalidateCache = invalidateCache || this.cacheInvalid() ) )
                 {this.invalidateCache();}
 
+                this.collectBbox( transform, drawableCollection, singlePath, invalidateCache, planeMask, clipPlanes );
+
                 // check if sub-graph can be culled away or render flag was set to false
                 planeMask = drawableCollection.cull( transform, this.graphState(), singlePath, planeMask );
                 if ( planeMask < 0 )